add multilevel static hash table

+++ src/com/company/Hasher.java
@@ -2,5 +2,5 @@ package com.company;
 
 public interface Hasher {
     void generateFunction();
-    long hash(long key);
+    int hash(int key);
 }
